+++ tests/test_emitter.py
@@ -70,11 +70,7 @@ def setup_teardown(tmpdir):
 
     yield
 
-    try:
-        emitter.stop()
-    except OSError:
-        # watch was already stopped, e.g., in `test_delete_self`
-        pass
+    emitter.stop()
     emitter.join(5)
     assert not emitter.is_alive()